inference/requirements.inference.txt
torch>=2.1
torchvision>=0.16
transformers>=4.38
tqdm>=4.66
numpy>=1.24
Pillow>=9.5
opencv-python>=4.8
openslide-python>=1.3
